added visualisation of the results

+++ src/main/java/pl/edu/pw/spdb/dal/DatabaseService.java
@@ -4,6 +4,6 @@ import pl.edu.pw.spdb.model.Point;
 import pl.edu.pw.spdb.model.Route;
 
 public interface DatabaseService {
-    Integer getStartOrEnd(Point point, boolean isStartPoint);
+    Long getStartOrEnd(Point point, boolean isStartPoint);
     Route findRoute(long startId, long endId, Integer maxSpeed, float distanceWeight);
 }
